requirements.txt
asgiref
certifi
crispy-bootstrap5
distlib
Django
django-cors-headers
django-crispy-forms
django-environ
django-filter
djangorestframework
filelock
Markdown
pillow
pipenv
platformdirs
pytz
setuptools
sqlparse
tzdata
virtualenv